Create OpenCV Coin Detection (Video and Image).py
The code makes allow users to choose any video or image file. Then it apply coin detection algorithms that can recognize and highlight the presence of circular objects that looks like coins. It shows the result by marking out the detected coins when an image is used, but when videos are used, it processes all frames in order to make a new output video with visualized coin detection while showing it also real-time. The python code excludes non-circular items and those which are smaller than typical coins.

1.	Library Used
•	OpenCV (cv2)
•	NumPy (np)
•	Tkinter
2.	Functions: 
•	Converts the input to grayscale
•	Applies blur to reduce noise
•	Uses adaptive thresholding as T values change w.r.t image
•	Filters contours based on area, circularity, and size to identify coins
•	Returns a list of detected coins 
3.	Process Image Functions: 
•	Calls detect_coins
•	Draws green circles (can be change tho)
4.	Process Video Function: 
•	Processes each frame
•	Draws green circles around detected coins in each frame
•	Creates an output video 
•	Displays in real-time